{"schema":"decodex/commit/1","summary":"Modularize run cycle dispatch","authority":"manual"}
